Implement SSR in Editor
