+++ testing/tests/DevExpress.serverSide/editor.markup.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/editor.markup.tests.js");
